backend/requirements.txt
fastapi==0.104.0
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
redis==5.0.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
pydantic==2.4.0
pydantic-settings==2.0.3
alembic==1.12.1
python-dotenv==1.0.0
aiofiles==23.2.1
bcrypt==4.1.2
requests==2.31.0
gunicorn==21.2.0